ref(normalization): Reduce processor module visibility
